fix(core): elements not being properly drawn after screen change
Since the widgets are reconfigured when bar is reconfigured, elements of
the widget are destroyed and constructed once more. This means, that
previous elements' state must now be passed on to newly created ones.
This patch ensures proper draw state even after a widget reconfiguration,
since all state is kept inside widget itself.